+++ src/components/20-molecules/dropdown/index.js
@@ -17,8 +17,9 @@ import { applyDefaults } from '../../../utils/with-react';
 const ARROW_ICON = svg([ExpandSvg]);
 const DEBOUNCE_DELAY = 250; // milliseconds
 const DROPDOWN_UL_MAXHEIGHT = '200px';
-const FOCUSENTRY_DIRECTION_DOWN = 'down';
-const FOCUSENTRY_DIRECTION_UP = 'up';
+const DOWN = '40'; // keyCode
+const UP = '38'; // keyCode
+const ARROW_KEY = { [UP]: -1, [DOWN]: 1 };
 
 // module globals
 let openDropdownInstance;
@@ -45,7 +46,7 @@ const forEach = (array, callback, scope) => {
   }
 };
 
-const nativeItemsMapper = ({ name, value, selected, disabled }, index) =>
+const nativeItemsMapper = ({ name, value, selected, _disabled }, index) =>
   html`
     <option
       class="m-dropdown__option"
@@ -53,27 +54,26 @@ const nativeItemsMapper = ({ name, value, selected, disabled }, index) =>
       data-index="${index}"
       data-value="${value}"
       ?selected="${selected}"
-      ?disabled="${disabled}"
+      ?disabled="${_disabled}"
       >${name}</option
     >
   `;
 
-const contentItemsMapper = (clickHandler, keyUpHandler, defaultTitle) => (
-  { name, value, selected, disabled },
+const contentItemsMapper = (clickHandler, defaultTitle) => (
+  { name, value, selected, _disabled },
   index
 ) => {
   const classes = {
     'm-dropdown__item': true,
     'm-dropdown__item--is-selected': selected,
   };
-  return disabled
+  return _disabled
     ? html``
     : html`
         <li class="${classMap(classes)}">
           <button
             type="button"
             @click="${clickHandler}"
-            @keyup="${keyUpHandler}"
             tabindex="-1"
             class="m-dropdown__button js-dropdown__button"
             data-index="${index + (defaultTitle ? 1 : 0)}"
@@ -87,7 +87,7 @@ const contentItemsMapper = (clickHandler, keyUpHandler, defaultTitle) => (
 
 const defaultTitleIfNeeded = (title, anotherSelection) =>
   title
-    ? [{ name: title, disabled: true, selected: !anotherSelection, value: '' }]
+    ? [{ name: title, _disabled: true, selected: !anotherSelection, value: '' }]
     : [];
 
 // CE
@@ -174,8 +174,16 @@ class AXADropdown extends NoShadowDOM {
     forEach(links, (_, link) =>
       link.setAttribute('tabindex', open ? '0' : '-1')
     );
-    if (open && openDropdownInstance) {
-      openDropdownInstance.open = false;
+    if (open) {
+      // close any other open instance in the same document
+      if (openDropdownInstance) {
+        openDropdownInstance.open = false;
+      }
+      // focus on selected item s.t. enhanced-mode keyboard navigation works like native
+      this.focusOnButton(
+        this.selectedIndex,
+        /* already includes default-title offset */ true
+      );
     }
     openDropdownInstance = open ? this : /* help GC */ null;
   }
@@ -186,52 +194,62 @@ class AXADropdown extends NoShadowDOM {
     }
   }
 
-  _focusEntry(direction) {
-    const { open, dropdown } = this;
+  focusOnButton(index, includesOffset) {
+    if (typeof index === 'number') {
+      const { defaultTitle, dropdown } = this;
+      const defaultTitleOffset = defaultTitle && !includesOffset ? 1 : 0;
+      dropdown
+        .querySelector(
+          `.js-dropdown__button[data-index="${index + defaultTitleOffset}"]`
+        )
+        .focus();
+    }
+  }
 
-    if (open) {
-      let selectedIndex = -1;
-      let newSelectedIndex;
-      const selectedLinkButton = dropdown.querySelector(
-        '.m-dropdown__item--is-selected > button'
-      );
-      const focusedLinkButton = dropdown.querySelector('button:focus');
-      const links = dropdown.querySelectorAll('.js-dropdown__button');
+  navigateByKeyboard(direction) {
+    const { items, defaultTitle } = this;
+    const { activeElement } = document;
+    const defaultTitleOffset = defaultTitle ? 1 : 0;
 
-      if (focusedLinkButton) {
-        forEach(links, (index, link) => {
-          if (link === focusedLinkButton) selectedIndex = index;
-        });
-      } else if (selectedLinkButton) {
-        forEach(links, (index, link) => {
-          if (link === selectedLinkButton) selectedIndex = index;
-        });
-      }
+    // don't get involved if preexisting focus outside enhanced-mode DOM
+    if (
+      !this.contains(activeElement) ||
+      !/(?:BUTTON|LI)/.test(activeElement.tagName)
+    ) {
+      return;
+    }
 
-      if (direction === FOCUSENTRY_DIRECTION_DOWN) {
-        newSelectedIndex = selectedIndex + 1;
-      } else {
-        newSelectedIndex = selectedIndex - 1;
-      }
+    // get selected item's index
+    let focussedIndex =
+      parseInt(activeElement.dataset.index || '0', 10) - defaultTitleOffset;
 
-      if (links[newSelectedIndex]) {
-        links[newSelectedIndex].focus();
-      }
+    // move in the direction of the arrow key
+    focussedIndex += direction;
+
+    // we landed on a focussable item?
+    if (items[focussedIndex] && !items[focussedIndex]._disabled) {
+      // yes, put the focus on it
+      this.focusOnButton(focussedIndex);
     }
   }
 
   handleKeyUp(e) {
     const { key, keyCode } = e;
     e.preventDefault();
-
+    const arrowKey = ARROW_KEY[keyCode.toString()];
+    // close dropdown via key?
     if (key === 'Escape' || key === 'Esc' || keyCode === 27) {
       this.openDropdown(false);
-    } else if (keyCode === 40) {
-      // arrowkey down
-      this._focusEntry(FOCUSENTRY_DIRECTION_DOWN);
-    } else if (keyCode === 38) {
-      // arrowkey up
-      this._focusEntry(FOCUSENTRY_DIRECTION_UP);
+    } else if (arrowKey) {
+      // arrow keys ↑, ↓,
+      // case 1: dropdown closed
+      if (!this.open) {
+        // open it in order to mimick native <select> behaviour
+        this.openDropdown(true);
+      } else {
+        // case 2: already open, navigate
+        this.navigateByKeyboard(arrowKey);
+      }
     }
   }
 
@@ -255,7 +273,7 @@ class AXADropdown extends NoShadowDOM {
         : target;
 
     const { value, index } = realTarget.dataset;
-    const { onChange, selectedIndex } = this;
+    const { onChange, selectedIndex, name } = this;
 
     this.openDropdown(false);
 
@@ -271,28 +289,42 @@ class AXADropdown extends NoShadowDOM {
     }
 
     this.selectedIndex = integerIndex;
-    const [{ name }] = this.findByValue(value);
+    const [{ name: optionLabel }] = this.findByValue(value);
     // allow idiomatic event.target.value in onChange callback!
-    const syntheticEvent = { target: { value, index: integerIndex, name } };
+    const details = {
+      value,
+      name,
+      optionLabel,
+      index: integerIndex,
+    };
+    const syntheticEvent = { target: details };
     onChange(syntheticEvent);
     if (!this.isControlled) {
       // declare the following value update to be uncontrolled
       this.state.firstTime = false;
       this.value = value; // triggers re-render
-      const details = { value, index: integerIndex, name };
       fireCustomEvent('axa-change', value, this, { bubbles: false });
       fireCustomEvent('change', details, this, { bubbles: false });
     }
   }
 
   findByValue(value, indexOnly) {
-    const { items = [] } = this;
+    const { items = [], defaultTitle } = this;
     const itemIndex = findIndex(items, ({ selected, value: selectedValue }) =>
       value === null
         ? selected
         : selectedValue === value || selectedValue === parseInt(value, 10)
     );
-    return indexOnly ? itemIndex : [items[itemIndex], itemIndex];
+    if (indexOnly) {
+      return itemIndex;
+    }
+
+    // no item selected, but defaultTitle set?
+    if (value === null && itemIndex < 0 && defaultTitle) {
+      return [false]; // signal caller (render) to use defaultTitle as title
+    }
+
+    return [items[itemIndex], itemIndex];
   }
 
   updateItems(value) {
@@ -366,7 +398,7 @@ class AXADropdown extends NoShadowDOM {
               : ''}
           </label>
         `}
-      <div class="m-dropdown__wrapper">
+      <div class="m-dropdown__wrapper" @keyup="${handleKeyUp}">
         <div class="m-dropdown__elements">
           <!-- NATIVE -->
           <div class="m-dropdown__select-wrapper">
@@ -398,7 +430,6 @@ class AXADropdown extends NoShadowDOM {
             @focus="${this.onFocus}"
             @blur="${this.onBlur}"
             @click="${handleDropdownClick}"
-            @keyup="${handleKeyUp}"
           >
             <span class="m-dropdown__flex-wrapper">
               <span class="js-dropdown__title">${this.title}</span>
@@ -413,11 +444,7 @@ class AXADropdown extends NoShadowDOM {
               : ''}"
           >
             ${items.map(
-              contentItemsMapper(
-                handleDropdownItemClick,
-                handleKeyUp,
-                defaultTitle
-              )
+              contentItemsMapper(handleDropdownItemClick, defaultTitle)
             )}
           </ul>
           <!-- ENHANCED END -->